fix: move test fixtures to autoload-dev, factory to tests/
Move Studio\Totem\Tests namespace from autoload to autoload-dev.
Move TotemUserFactory from database/factories/ to tests/ under the
Studio\Totem\Tests namespace, keeping test-only code out of the
production classmap.
