+++ composer.json
@@ -40,13 +40,17 @@
   "autoload": {
     "psr-4": {
       "Studio\\Totem\\": "src/",
-      "Studio\\Totem\\Tests\\": "tests/",
       "Database\\Factories\\": "database/factories/",
       "Database\\Seeders\\": "database/seeders/"
     },
     "files": [
       "src/helpers.php"
     ]
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "Studio\\Totem\\Tests\\": "tests/"
+    }
   },
   "extra": {
     "component": "package",
