+++ tests/TestUser.php
@@ -2,7 +2,6 @@
 
 namespace Studio\Totem\Tests;
 
-use Database\Factories\TotemUserFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
